+++ src/program-types.ts
@@ -6,7 +6,6 @@ import * as anchor from "@project-serum/anchor";
 export interface State {
   admin: PublicKey;
   stateNonce: number;
-  vaultNonce: number;
   serumNonce: number;
   mintAuthNonce: number;
   numUnderlyings: number;
@@ -73,6 +72,12 @@ export interface MarginParameters {
   padding: Array<number>;
 }
 
+export interface PerpParameters {
+  minFundingRatePercent: anchor.BN;
+  maxFundingRatePercent: anchor.BN;
+  impactCashDelta: anchor.BN;
+}
+
 export interface HaltState {
   halted: boolean;
   spotPrice: anchor.BN;
@@ -98,12 +103,15 @@ export interface ZetaGroup {
   marginParameters: MarginParameters;
   products: Array<Product>;
   productsPadding: Array<Product>;
+  perp: Product;
   expirySeries: Array<ExpirySeries>;
   expirySeriesPadding: Array<ExpirySeries>;
   totalInsuranceVaultDeposits: anchor.BN;
   asset: any;
   expiryIntervalSeconds: number;
   newExpiryThresholdSeconds: number;
+  perpParameters: PerpParameters;
+  perpSyncQueue: PublicKey;
   padding: Array<number>;
 }
 
@@ -153,9 +161,11 @@ export interface MarginAccount {
   seriesExpiry: Array<anchor.BN>;
   productLedgers: Array<ProductLedger>;
   productLedgersPadding: Array<ProductLedger>;
+  perpProductLedger: ProductLedger;
   rebalanceAmount: anchor.BN;
   asset: any;
   accountType: any;
+  lastFundingDelta: AnchorDecimal;
   padding: Array<number>;
 }
 
@@ -164,6 +174,7 @@ export interface SpreadAccount {
   nonce: number;
   balance: anchor.BN;
   seriesExpiry: Array<anchor.BN>;
+  seriesExpiryPadding: anchor.BN;
   positions: Array<Position>;
   positionsPadding: Array<Position>;
   asset: any;
@@ -173,6 +184,7 @@ export interface SpreadAccount {
 export interface Greeks {
   nonce: number;
   markPrices: Array<anchor.BN>;
+  perpMarkPrice: anchor.BN;
   markPricesPadding: Array<anchor.BN>;
   productGreeks: Array<ProductGreeks>;
   productGreeksPadding: Array<ProductGreeks>;
@@ -187,7 +199,18 @@ export interface Greeks {
   volatilityPadding: Array<anchor.BN>;
   nodeKeys: Array<PublicKey>;
   haltForcePricing: Array<boolean>;
+  perpUpdateTimestamp: anchor.BN;
+  perpFundingDelta: AnchorDecimal;
+  perpLatestFundingRate: AnchorDecimal;
+  perpLatestMidpoint: anchor.BN;
   padding: Array<number>;
+}
+
+export interface PerpSyncQueue {
+  nonce: number;
+  head: number;
+  length: number;
+  queue: Array<AnchorDecimal>;
 }
 
 export interface MarketNode {
@@ -318,4 +341,14 @@ export interface OrderCompleteEvent {
   orderId: anchor.BN;
   clientOrderId: anchor.BN;
   orderCompleteType: Object;
+}
+
+export interface ApplyFundingEvent {
+  marginAccount: PublicKey;
+  user: PublicKey;
+  asset: Object;
+  balanceChange: anchor.BN;
+  remainingBalance: anchor.BN;
+  fundingRate: anchor.BN;
+  oraclePrice: anchor.BN;
 }